LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-response-xhr.https-expected.txt


FAIL Verify the response of FetchEvent using XMLHttpRequest assert_equals: expected (string) "foo, bar" but got (object) null
